syncing components/contexts (extracting from components/dataset_base.py)
