+++ tests/unit/test_dataset.py
@@ -18,15 +18,13 @@ from datumaro.components.annotation import (
     Polygon,
     PolyLine,
 )
-from datumaro.components.dataset import DEFAULT_FORMAT, Dataset, eager_mode
-from datumaro.components.dataset_base import (
-    DatasetBase,
-    DatasetItem,
+from datumaro.components.contexts.importer import (
     FailingImportErrorPolicy,
     ImportErrorPolicy,
     ProgressReporter,
-    SubsetBase,
 )
+from datumaro.components.dataset import DEFAULT_FORMAT, Dataset, eager_mode
+from datumaro.components.dataset_base import DatasetBase, DatasetItem, SubsetBase
 from datumaro.components.dataset_item_storage import ItemStatus
 from datumaro.components.environment import Environment
 from datumaro.components.errors import (
